UX: Multichain: Send Flow: Move connected components to individual components
