+++ ui/components/multichain/pages/send/components/index.js
@@ -1,0 +1,4 @@
+export { SendPageRow } from './send-page-row';
+export { SendPageAccountPicker } from './account-picker';
+export { SendPageNetworkPicker } from './network-picker';
+export { SendPageYourAccount } from './your-accounts';
